update Jekyll
Using the version of ruby that is supplied by apt-get on Ubuntu [1], I've updated
the version of Jekyll that is used to build the site.

$ bundle install
[...]
ruby_dep-1.5.0 requires ruby version >= 2.2.5, ~> 2.2, which is incompatible with the current version, ruby 3.0.2p107

$ bundle update
[...]
Bundle updated!

$ bundle add webrick
[...]

And finally, building and running the site works again:
$ bundle exec jekyll serve
[...]
Server running...

[1] ruby 3.0.2p107 (2021-07-07 revision 0db68f0233) [x86_64-linux-gnu]